refactor: own hook-run provenance reads in the validation domain (#145)
The HookRunReader port I'd put in the feature domain was wrong: it inverted
ownership (feature dictating a validation-aggregate query) and laundered a real
feature->validation dependency through an infra adapter that read validation's
tables.

hook_runs/hook_releases are validation-domain data, so the read capability
belongs there: run_ids_for_batch / run_ids_for_deposition now live on the
HookRegistry port + Postgres adapter (SQL) + HookRegistryService. The feature
insert handlers call the validation service directly — an honest, declared
service-to-service cross-domain read. Deleted the feature-domain reader port,
its adapter, and the DI binding.
