+++ server/osa/domain/feature/handler/insert_record_features.py
@@ -1,10 +1,10 @@
 """InsertRecordFeatures — deferred feature insertion on record publication."""
 
-from osa.domain.feature.port.hook_run_reader import HookRunReader
 from osa.domain.feature.port.storage import FeatureStoragePort
 from osa.domain.feature.service.feature import FeatureService
 from osa.domain.record.event.record_published import RecordPublished
 from osa.domain.shared.event import EventHandler
+from osa.domain.validation.service.hook_registry import HookRegistryService
 
 
 class InsertRecordFeatures(EventHandler[RecordPublished]):
@@ -16,7 +16,7 @@ class InsertRecordFeatures(EventHandler[RecordPublished]):
 
     feature_service: FeatureService
     feature_storage: FeatureStoragePort
-    hook_run_reader: HookRunReader
+    hook_registry: HookRegistryService
 
     async def handle(self, event: RecordPublished) -> None:
         hook_output_dir = self.feature_storage.get_hook_output_root(
@@ -24,7 +24,7 @@ class InsertRecordFeatures(EventHandler[RecordPublished]):
         )
         # Deposition records carry the deposition SRN as their source id; the
         # hook_runs were recorded against it, so resolve {hook_name: run_id}.
-        run_ids = await self.hook_run_reader.run_ids_for_deposition(event.source.id)
+        run_ids = await self.hook_registry.run_ids_for_deposition(event.source.id)
         await self.feature_service.insert_features_for_record(
             hook_output_dir=hook_output_dir,
             record_srn=str(event.record_srn),
