+++ server/osa/infrastructure/persistence/di.py
@@ -28,7 +28,6 @@ from osa.domain.shared.outbox import Outbox
 from osa.domain.shared.port.event_repository import EventRepository
 from osa.domain.feature.port.feature_store import FeatureStore
 from osa.domain.validation.port.repository import ValidationRunRepository
-from osa.domain.feature.port.hook_run_reader import HookRunReader
 from osa.domain.validation.port.hook_registry import HookRegistry
 from osa.domain.data.port.data_read_store import (
     DataCatalogReadStore,
@@ -50,9 +49,6 @@ from osa.infrastructure.persistence.repository.convention import (
 )
 from osa.infrastructure.persistence.repository.hook_registry import (
     PostgresHookRegistry,
-)
-from osa.infrastructure.persistence.repository.hook_run_reader import (
-    PostgresHookRunReader,
 )
 from osa.infrastructure.persistence.repository.deposition import (
     PostgresDepositionRepository,
@@ -132,11 +128,9 @@ class PersistenceProvider(Provider):
         PostgresConventionRepository, scope=Scope.UOW, provides=ConventionRepository
     )
 
-    # Hook registry (validation domain — feature #145)
+    # Hook registry (validation domain — feature #145). Owns hooks, releases,
+    # the live pointer, and hook_runs (record + provenance reads).
     hook_registry_repo = provide(PostgresHookRegistry, scope=Scope.UOW, provides=HookRegistry)
-
-    # Provenance read: resolve {hook_name: run_id} for a batch/deposition.
-    hook_run_reader = provide(PostgresHookRunReader, scope=Scope.UOW, provides=HookRunReader)
 
     # Cross-domain readers
     schema_reader = provide(SchemaReaderAdapter, scope=Scope.UOW, provides=SchemaReader)
